third_party/blink/web_tests/platform/generic/http/tests/websocket/send-onmessage-origin-expected.txt

WebSocket message event origin attribute test (String message)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
PASS data is firstMessageToSend
PASS wsOrigin is "ws://localhost:8880"
PASS data is secondMessageToSend
PASS wsOrigin is "ws://localhost:8880"
Closed.
PASS successfullyParsed is true

TEST COMPLETE
